refactor(async-jobs): extract startAsyncJob/pruneAsyncJobs/toPublicJob into async-jobs.js
Introduces createAsyncJobManager({ db, asyncJobs, ttlMs, runnerDir }) factory
that encapsulates the four async job functions previously defined inline in
index.js:
- startAsyncJob: child process spawn, progress parsing, checkpoint writes
- pruneAsyncJobs: TTL eviction, tmpDir cleanup, DB pruning
- toPublicJob: job → public shape serialization
- readJsonIfExists: exported as a standalone utility (also used by index.js
  for package.json version reading)

index.js removes os, spawn, and ASYNC_PROGRESS_PREFIX imports; wires the
factory after db open and stalled-job recovery. asyncJobs Map and
waitForRunningJobs remain in index.js (shutdown drain needs direct Map access).
No behavioral changes.
